# services/product-service/src/main/java/org/openapitools/api/ProductApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.CreateProduct201Response;
import org.openapitools.model.Error;
import org.openapitools.model.Product;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-14T11:50:51.562192-07:00[America/Los_Angeles]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "Product", description = "the Product API")
public interface ProductApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_PRODUCT = "/product";
    /**
     * POST /product : Create a product, with the provided details
     * Create a product, with the provided details. Server creates the new product_id.
     *
     * @param product  (required)
     * @return Product created successfully (status code 201)
     *         or Invalid input data (status code 400)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Create a product, with the provided details",
        description = "Create a product, with the provided details. Server creates the new product_id.",
        tags = { "Product" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Product created successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CreateProduct201Response.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid input data", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ProductApi.PATH_CREATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<CreateProduct201Response> createProduct(
        @Parameter(name = "Product", description = "", required = true) @Valid @RequestBody Product product
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"product_id\" : 0 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PRODUCT = "/products/{productId}";
    /**
     * GET /products/{productId} : Get product by ID
     * Retrieve a product&#39;s details using its unique identifier
     *
     * @param productId Unique identifier for the product (required)
     * @return Product found successfully (status code 200)
     *         or Product not found (status code 404)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "getProduct",
        summary = "Get product by ID",
        description = "Retrieve a product's details using its unique identifier",
        tags = { "Product" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Product found successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Product.class))
            }),
            @ApiResponse(responseCode = "404", description = "Product not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductApi.PATH_GET_PRODUCT,
        produces = { "application/json" }
    )
    default ResponseEntity<Product> getProduct(
        @NotNull @Min(value = 1) @Parameter(name = "productId", description = "Unique identifier for the product", required = true, in = ParameterIn.PATH) @PathVariable("productId") Integer productId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"category_id\" : 456, \"some_other_id\" : 789, \"product_id\" : 12345, \"weight\" : 1250, \"sku\" : \"ABC123XYZ\", \"manufacturer\" : \"Acme Corporation\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}