# services/shopping-cart-service/src/main/java/org/openapitools/api/PaymentsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.Error;
import org.openapitools.model.ProcessPaymentRequest;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-14T11:50:52.847186-07:00[America/Los_Angeles]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "Payments", description = "Payment processing operations")
public interface PaymentsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_PROCESS_PAYMENT = "/credit-card-authorizer/authorize";
    /**
     * POST /credit-card-authorizer/authorize : Authorize and process credit card payment, called by shoppinCart.checkout. Internal API
     * Process payment for a shopping cart using credit card information
     *
     * @param processPaymentRequest  (required)
     * @return Payment authorized successfully (status code 200)
     *         or Invalid payment information (status code 400)
     *         or Payment declined (status code 402)
     *         or Internal server error (status code 500)
     */
    @Operation(
        operationId = "processPayment",
        summary = "Authorize and process credit card payment, called by shoppinCart.checkout. Internal API",
        description = "Process payment for a shopping cart using credit card information",
        tags = { "Payments" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Payment authorized successfully"),
            @ApiResponse(responseCode = "400", description = "Invalid payment information", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "402", description = "Payment declined", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth"),
            @SecurityRequirement(name = "BearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = PaymentsApi.PATH_PROCESS_PAYMENT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> processPayment(
        @Parameter(name = "ProcessPaymentRequest", description = "", required = true) @Valid @RequestBody ProcessPaymentRequest processPaymentRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"details\" : \"Product ID must be a positive integer\", \"error\" : \"INVALID_INPUT\", \"message\" : \"The provided input data is invalid\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}